- added CigaretteDetailsWidget for actual details of cigarette to input into database. - added CigaretteDetailsWidget route mapping to routes.dart.
